mysummit command fix for multi platform

+++ LiveBot3/DB/CommandsUsedCount.cs
@@ -8,7 +8,7 @@ namespace LiveBot.DB
     {
         [Key]
         [Column("command_id")]
-        public long Command_ID { get; set; }
+        public int Command_ID { get; set; }
 
         [Required]
         [Column("command")]
